+++ opencc_purepy/union_cache.py
@@ -17,6 +17,7 @@ class UnionKey(Enum):
     # TW helpers
     TwPhrasesOnly = auto()
     TwVariantsPair = auto()
+    S2TwpR2TwTriple = auto()
     TwPhrasesRevOnly = auto()
     TwRevPair = auto()
     Tw2SpR1TwRevTriple = auto()
@@ -93,6 +94,8 @@ class UnionCache:
             return [g("tw_phrases")]
         if key is UnionKey.TwVariantsPair:
             return [g("tw_variants_phrases"), g("tw_variants")]
+        if key is UnionKey.S2TwpR2TwTriple:
+            return [g("tw_phrases"), g("tw_variants_phrases"), g("tw_variants")]
         if key is UnionKey.TwPhrasesRevOnly:
             return [g("tw_phrases_rev")]
         if key is UnionKey.TwRevPair: